% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{add_layer}
\alias{add_layer}
\title{Add Layer for hierarchical dataframe}
\usage{
add_layer(
  .data,
  child_col,
  node_type_col = NULL,
  node_type_vals = " ",
  link_col = NULL,
  link_vals = " ",
  attribute_cols = NULL
)
}
\arguments{
\item{.data}{Output from \code{add_root()} or \code{add_layer()}.}

\item{child_col}{Column to generate children from.}

\item{node_type_col, node_type_vals}{Column or values to use as node_type.
If not specified, all node types will be assigned as blank (" ") for the layer.}

\item{link_col, link_vals}{Column or values to use as link.
If not specified, all links will be assigned as blank (" ") for the layer.}

\item{attributes}{}
}
\description{
Used in conjunction with \code{add_root()} to generate dataframe for use with
\code{hp_dataframe()}.
}
\examples{
os_survey \%>\%
  add_root("OS Students 2014/15") \%>\%
  add_layer(
    child_col = "Operating System",
    link_vals = "OS",
    node_type_vals = "OS"
  ) \%>\%
  add_layer(
    child_col = "OS Version",
    link_vals = "Ver",
    node_type_vals = "Sub",
    attribute_cols = "users"
  ) \%>\%
  hp_dataframe(
    title = "Survey Results of Most Popular OS in 2014/15",
    settings = hierplane_settings(attributes = "attribute1"),
    styles = hierplane_styles(
      link_to_positions = list(Ver = "right")
    )
  ) \%>\%
  hierplane()

}
