% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{add_root}
\alias{add_root}
\title{Add Root for hierarchical dataframe}
\usage{
add_root(.data, root, attribute = NULL)
}
\arguments{
\item{.data}{Dataframe to construct hierarchical dataframe from (source data).}

\item{root}{Value to display as root.}

\item{attribute}{Values to tag root (i.e. top level) plane with.}
}
\description{
Lays foundation for hierarchical dataframe used for use with \code{hp_dataframe()}.
}
\examples{
os_survey \%>\% add_root("OS Students 2014/15")

}
