% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_styles.R
\name{add_styles}
\alias{add_styles}
\title{Add styles to hierplane settings}
\usage{
add_styles(
  settings,
  node_type_to_style = NULL,
  link_to_positions = NULL,
  link_name_to_label = NULL
)
}
\arguments{
\item{settings}{A list of settings from \code{hierplane_settings()}}

\item{node_type_to_style}{TODO}

\item{link_to_positions}{TODO}

\item{link_name_to_label}{TODO}
}
\description{
Append styling options to output of \code{hierplane_settings()}.
}
