% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{build_nodes}
\alias{build_nodes}
\title{Build Hierarchical Tree Nodes}
\usage{
build_nodes(
  sp_tib,
  root = sp_tib[1, 1],
  parent_id = "head_token_id",
  parent = "head_token",
  nodetype = "dep_",
  link = "dep_",
  attributes = c("ent_type_", "pos", "is_currency", "like_url", "like_email"),
  word = "token",
  word_id = "token_id"
)
}
\arguments{
\item{sp_tib}{Tibble generated from \code{spacy_tibble}.}

\item{root}{Reference to first headtoken in sp_tib.}

\item{nodetype, word, link, attributes, word_id}{Reference to fields used to populate node.}
}
\description{
Build children nodes.
https://stackoverflow.com/questions/23839142/transform-a-dataframe-into-a-tree-structure-list-of-lists
}
