% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{build_tree}
\alias{build_tree}
\title{Build hierarchical tree}
\usage{
build_tree(
  txt,
  nodetype = "dep_",
  word = "token",
  link = "dep_",
  attributes = c("ent_type_", "pos", "is_currency", "like_url", "like_email"),
  word_id = "token_id"
)
}
\value{
A list containing tree in list and JSON formats.
}
\description{
Wrapper for \code{build_node} to Recursively build hierarchical tree from children nodes.
}
