% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierplane.R
\name{hierplane}
\alias{hierplane}
\title{Render a hierplane}
\source{
https://github.com/allenai/hierplane
}
\usage{
hierplane(
  .data,
  theme = "light",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{.data}{A dataframe containing hierarchical features.}

\item{theme}{Either light or dark}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Explicitly provide an element ID to the widget rather than
an automatically generated one. This is optional.}
}
\description{
Hierplanes visualize hierarchical data. The \code{.data} param expects a hierplane
object created by various inputs using the \code{hp_} family of functions.
}
