% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierplane_settings.R
\name{hierplane_settings}
\alias{hierplane_settings}
\title{Assign column names to hierplane}
\usage{
hierplane_settings(
  type = "hier",
  parent_id = "parent_id",
  child_id = "child_id",
  child = "child",
  node_type = "node_type",
  link = "link",
  root_tag = "ROOT",
  attributes = c("attribute1", "attribute2")
)
}
\arguments{
\item{type}{Either \code{hier} for standard hierplane or \code{spacy} for a spacy hierplane.}

\item{parent_id}{TODO}

\item{child_id}{TODO}

\item{child}{TODO}

\item{node_type}{TODO}

\item{link}{TODO}

\item{root_tag}{TODO}

\item{attributes}{TODO}
}
\description{
Hierplane requires a specific data structure with things like parent IDs,
child IDs, the root ID, etc. This function requires the user to identify
which columns in their dataframe represent each of these requirements.
}
