% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy.R
\name{hierplane_spacy}
\alias{hierplane_spacy}
\title{Render a Hierplane from Text using Spacy}
\usage{
hierplane_spacy(txt, settings = spacy_default(), ...)
}
\arguments{
\item{txt}{A sentence to be parsed and rendered as hierplane}

\item{settings}{A hierplane settings object.}

\item{...}{Parameters to pass to \code{hierplane()}}
}
\description{
Render a Hierplane from Text using Spacy
}
