% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierplane_styles.R
\name{hierplane_styles}
\alias{hierplane_styles}
\title{Add styling options}
\usage{
hierplane_styles(
  node_type_to_style = NULL,
  link_to_positions = NULL,
  link_name_to_label = NULL
)
}
\arguments{
\item{node_type_to_style}{A named list of mappings from node type to styles.
See \code{hierplane::style_options$node_type_to_style} for available options.}

\item{link_to_positions}{A named list of mappings from link to node placements.
See \code{hierplane::style_options$link_to_positions} for available options.}

\item{link_name_to_label}{A named list of mappings from link names to labels (e.g. from "Episode" to "Ep").}
}
\description{
Add styling options
}
