% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_dataframe.R
\name{hp_dataframe}
\alias{hp_dataframe}
\title{Create a hierplane object from dataframe}
\usage{
hp_dataframe(.data, settings, title = "Hierplane")
}
\arguments{
\item{.data}{A dataframe with valid hierachical features (child_id, parent_id, etc.).}

\item{settings}{Assign your dataframes columns to hierplane attributes, additionally apply rendering styles.}

\item{title}{A title, defaults to "Hierplane", this serves as the header/title of the hierplane.}
}
\description{
Creating a hierplane object from a dataframe requires a bit more work because
we need input from the user to construct valid JSON. Construction requires the
user to provide a dataframe, title, and a list of settings.
}
