% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_datatree.R
\name{hp_datatree}
\alias{hp_datatree}
\title{Create a hierplane object using data.tree}
\usage{
hp_datatree(
  .data,
  title = "Hierplane",
  attributes = NULL,
  link = "to",
  styles = NULL
)
}
\arguments{
\item{.data}{A \code{data.tree} object of class "Node".}

\item{title}{A title, defaults to "Hierplane", this serves as the header/title of the hierplane.}

\item{attributes}{Attributes to assign to the nodes, these are the annotation in the nodes.}

\item{link}{Link connecting each node, theres are the tabs or connections you see between each node.}

\item{styles}{Assign styles to hierplane generated from \code{hierplane_styles()}.}
}
\description{
The \code{data.tree} package is a popular, general purpose hierarchical data
structure for R. Therefore, \code{hp_datatree} tries to make it so hierplane is
compatible with \code{data.tree} objects.
}
\examples{
\dontrun{
library(data.tree)
library(yaml)

"
name: r4fun
tyler:
  name: Tyler
  job: Data Scientist
  species: Human
  toulouse:
    name: Toulouse
    job: Systems Engineer
    species: Cat
    jojo:
      name: Jojo
      job: Python Programmer
      species: Cat
  ollie:
    name: Ollie
    job: Database Administrator
    species: Dog
  lucas:
    name: Lucas
    job: R Programmer
    species: Rabbit
" -> yaml

yaml \%>\%
  yaml.load() \%>\%
  as.Node() \%>\%
  hp_datatree(
    title = "r4fun github group",
    link = "species",
    attributes = "job"
  ) \%>\%
  hierplane(
    theme = "light",
    width = "auto",
    height = "auto"
  )
}
}
