% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hp_starwars}
\alias{hp_starwars}
\title{Starwars Films and Starships}
\format{
An object of class \code{data.frame} with 73 rows and 11 columns.
}
\usage{
hp_starwars
}
\description{
A sample hierplane dataframe derived from the {dplyr} \code{starwars} dataset that
shows the appearances of each starship in the first seven Star Wars movies.
}
\details{
Source: \url{https://dplyr.tidyverse.org/reference/starwars.html}
}
\keyword{datasets}
