% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{os_survey}
\alias{os_survey}
\title{Operating System Survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 4 columns.
}
\usage{
os_survey
}
\description{
A dataframe showing popularity of operating systems among students.
}
\details{
Source: \url{https://cran.r-project.org/web/packages/data.tree/vignettes/data.tree.html}
}
\keyword{datasets}
