% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_spacyr.R
\name{spacyr_default_attributes}
\alias{spacyr_default_attributes}
\title{Default settings for hp_spacyr attributes}
\usage{
spacyr_default_attributes()
}
\description{
Default settings for hp_spacyr attributes
}
