% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierplane_styles.R
\docType{data}
\name{style_options}
\alias{style_options}
\title{List of all available styles}
\format{
An object of class \code{list} of length 3.
}
\usage{
style_options
}
\description{
A list containing all available styling options for hierplane
}
\details{
Source: \url{https://github.com/allenai/hierplane/}
}
\keyword{datasets}
