% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierplane_settings.R
\name{hierplane_settings}
\alias{hierplane_settings}
\title{Assign column names to hierplane}
\usage{
hierplane_settings(
  type = "hier",
  parent_id = "parent_id",
  child_id = "child_id",
  child = "child",
  node_type = "node_type",
  link = "link",
  root_tag = "ROOT",
  attributes = NULL
)
}
\arguments{
\item{type}{Either \code{hier} for standard hierplane or \code{spacy} for a spacy hierplane.}

\item{parent_id}{Column to be used for linkage references when generating children nodes.}

\item{child_id}{Column to be used for \emph{generating} children nodes.}

\item{child}{Column to be used for \emph{labeling} children nodes.}

\item{node_type}{Column to be used as node type (used for plane styling).}

\item{link}{Column to be used to generate lane tags (i.e. connections between planes).}

\item{root_tag}{Keyword in \code{link} field to be used to identify top-level plane. Defaulted to "ROOT".}

\item{attributes}{Column(s) to be used for generating the attribute tags.
If not specified (i.e. NULL), all columns with "attribute" in the name will be used.}
}
\description{
Hierplane requires a specific data structure with things like parent IDs,
child IDs, the root ID, etc. This function requires the user to identify
which columns in their dataframe represent each of these requirements.
}
