% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_dataframe.R
\name{hp_dataframe}
\alias{hp_dataframe}
\title{Create a hierplane object from dataframe}
\usage{
hp_dataframe(
  .data,
  title = "Hierplane",
  parent_id = "parent_id",
  child_id = "child_id",
  child = "child",
  node_type = "node_type",
  link = "link",
  root_tag = "ROOT",
  attributes = NULL,
  styles = NULL
)
}
\arguments{
\item{.data}{A dataframe with valid hierachical features (child_id, parent_id, etc.).}

\item{title}{A title, defaults to "Hierplane", this serves as the header/title of the hierplane.}

\item{parent_id}{Variable to be used for linkage references when generating children nodes.}

\item{child_id}{Variable to be used for \emph{generating} children nodes.}

\item{child}{Variable to be used for \emph{labeling} children nodes.}

\item{node_type}{Variable to be used as node type (used for plane styling).}

\item{link}{Variable to be used to generate lane tags (i.e. connections between planes).}

\item{root_tag}{Keyword in \code{link} field to be used to identify top-level plane. Defaulted to "ROOT".}

\item{attributes}{Variable(s) to be used for generating the attribute tags.
If not specified (i.e. NULL), all variables with "attribute" in the name will be used.}

\item{styles}{Assign styles to hierplane generated from \code{hierplane_styles()}.}
}
\description{
Creating a hierplane object from a dataframe requires a bit more work because
we need input from the user to construct valid JSON. Construction requires the
user to provide a dataframe, title, and a list of settings.
}
